NSInteger isn't defined in the tiger frameworks, which we use in our includes for dep-debug. Doesn't break [21372]

+++ Frameworks/AIUtilities Framework/Source/AIImageViewWithImagePicker.m
@@ -536,7 +536,7 @@
 	[self showPictureTaker];
 }
 
-- (void)pictureTakerDidEnd:(id)inPictureTaker returnCode:(NSInteger)returnCode contextInfo:(void *)contextInfo
+- (void)pictureTakerDidEnd:(id)inPictureTaker returnCode:(int)returnCode contextInfo:(void *)contextInfo
 {	
 	if (returnCode == NSOKButton) {
 		NSImage *image = [inPictureTaker outputImage];
